requirements.txt
django==4.2.10
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
coreapi==2.3.3
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langchain-openai==0.0.7
qdrant-client==1.7.3
python-docx==1.1.0
